# publisher/internal/service/gen/http/fileserver/server/encode_decode.go
// Code generated by goa v3.19.1, DO NOT EDIT.
//
// fileserver HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/PingCAP-QE/ee-apps/publisher/internal/service/design -o
// ./service

package server

import (
	"context"
	"errors"
	"io"
	"net/http"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeRequestToPublishResponse returns an encoder for responses returned by
// the fileserver request-to-publish endpoint.
func EncodeRequestToPublishResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeRequestToPublishRequest returns a decoder for requests sent to the
// fileserver request-to-publish endpoint.
func DecodeRequestToPublishRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			body RequestToPublishRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateRequestToPublishRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewRequestToPublishPayload(&body)

		return payload, nil
	}
}

// EncodeQueryPublishingStatusResponse returns an encoder for responses
// returned by the fileserver query-publishing-status endpoint.
func EncodeQueryPublishingStatusResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeQueryPublishingStatusRequest returns a decoder for requests sent to
// the fileserver query-publishing-status endpoint.
func DecodeQueryPublishingStatusRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			requestID string

			params = mux.Vars(r)
		)
		requestID = params["request_id"]
		payload := NewQueryPublishingStatusPayload(requestID)

		return payload, nil
	}
}