# publisher/internal/service/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Publish API","description":"Publish API","contact":{"name":"WuHui Zuo","url":"https://github.com/wuhuizuo","email":"wuhui.zuo@pingcap.com"},"version":"1.0.0"},"servers":[{"url":"http://0.0.0.0:80"}],"paths":{"/fs/publish-request":{"post":{"tags":["fileserver"],"summary":"request-to-publish fileserver","operationId":"fileserver#request-to-publish","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestToPublishRequestBody2"},"example":{"artifact_url":"Tenetur mollitia consequatur perferendis."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"type":"string","example":"Debitis eligendi iusto odio eveniet est."},"description":"request track ids","example":["Aperiam facere.","Vitae consequatur.","Inventore saepe magnam neque ut.","Tempore aut pariatur."]},"example":["Laboriosam nihil aut facere officiis est.","Doloremque hic.","Dolor odio corrupti sunt velit voluptas autem.","Cupiditate deleniti et dignissimos."]}}}}}},"/fs/publish-request/{request_id}":{"get":{"tags":["fileserver"],"summary":"query-publishing-status fileserver","operationId":"fileserver#query-publishing-status","parameters":[{"name":"request_id","in":"path","description":"request track id","required":true,"schema":{"type":"string","description":"request track id","example":"Minima voluptates."},"example":"Ea enim sint."}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","description":"request state","example":"processing","enum":["queued","processing","success","failed","canceled"]},"example":"queued"}}}}}},"/image/copy":{"post":{"tags":["image"],"summary":"request-to-copy image","operationId":"image#request-to-copy","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestToCopyRequestBody"},"example":{"destination":"Quia sed sunt vero.","source":"Iusto ut et deserunt nisi voluptates."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","description":"request id","example":"38844f96-6b37-405c-83e2-4bb2b8ab9280","format":"uuid"},"example":"ac9e55f2-f551-41ea-8b02-a600f46484f9"}}}}}},"/image/copy/{request_id}":{"get":{"tags":["image"],"summary":"query-copying-status image","operationId":"image#query-copying-status","parameters":[{"name":"request_id","in":"path","description":"request track id","required":true,"schema":{"type":"string","description":"request track id","example":"fdb5a585-2e5d-4563-afa4-78093f127433","format":"uuid"},"example":"ae97cf87-9621-4c3e-96db-c8c920d9f986"}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","description":"request state","example":"success","enum":["queued","processing","success","failed","canceled"]},"example":"failed"}}}}}},"/tiup/publish-request":{"post":{"tags":["tiup"],"summary":"request-to-publish tiup","operationId":"tiup#request-to-publish","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestToPublishRequestBody"},"example":{"artifact_url":"Consequatur ea eum saepe.","request_id":"Voluptates voluptatem accusamus nisi omnis quia molestias.","tiup-mirror":"Voluptas dolore eos eveniet vero voluptas.","version":"Omnis dolorem eveniet fugit nemo."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"type":"string","example":"Earum nemo ut autem molestias."},"description":"request track ids","example":["Omnis quasi repudiandae et omnis nesciunt.","Et dolorem odio quos voluptatem."]},"example":["Nostrum odio molestiae ab quis.","Quasi est quia.","Tempora consequatur.","Exercitationem nostrum neque quo atque et."]}}}}}},"/tiup/publish-request/{request_id}":{"get":{"tags":["tiup"],"summary":"query-publishing-status tiup","operationId":"tiup#query-publishing-status","parameters":[{"name":"request_id","in":"path","description":"request track id","required":true,"schema":{"type":"string","description":"request track id","example":"Doloremque ea rerum alias excepturi facilis ab."},"example":"Modi quia iste."}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","description":"request state","example":"processing","enum":["queued","processing","success","failed","canceled"]},"example":"success"}}}}}},"/tiup/reset-rate-limit":{"post":{"tags":["tiup"],"summary":"reset-rate-limit tiup","operationId":"tiup#reset-rate-limit","responses":{"200":{"description":"OK response."}}}}},"components":{"schemas":{"RequestToCopyRequestBody":{"type":"object","properties":{"destination":{"type":"string","description":"destination image url","example":"Est suscipit."},"source":{"type":"string","description":"source image url","example":"Voluptatibus porro modi eum ut."}},"example":{"destination":"Animi et sunt quidem.","source":"Excepturi voluptatem et voluptatum doloribus ratione."},"required":["source","destination"]},"RequestToPublishRequestBody":{"type":"object","properties":{"artifact_url":{"type":"string","description":"The full url of the pushed OCI artifact, contain the tag part. It will parse the repo from it.","example":"Deleniti repellat et et vel eaque qui."},"request_id":{"type":"string","description":"The request id","example":"Nemo et in eaque repellendus."},"tiup-mirror":{"type":"string","description":"Staging is http://tiup.pingcap.net:8988, product is http://tiup.pingcap.net:8987.","default":"http://tiup.pingcap.net:8988","example":"Sit commodi rem eum aut omnis."},"version":{"type":"string","description":"Force set the version. Default is the artifact version read from `org.opencontainers.image.version` of the manifest config.","example":"Aut dolore recusandae distinctio suscipit tempora."}},"example":{"artifact_url":"Veniam velit vero eos itaque saepe provident.","request_id":"Asperiores rerum est dolorem.","tiup-mirror":"Qui velit consequatur.","version":"Vitae non quidem provident qui recusandae dolor."},"required":["artifact_url","tiup-mirror"]},"RequestToPublishRequestBody2":{"type":"object","properties":{"artifact_url":{"type":"string","description":"The full url of the pushed OCI artifact, contain the tag part. It will parse the repo from it.","example":"Deserunt ut atque modi ipsum."}},"example":{"artifact_url":"Quis veritatis ut rerum quos consectetur."},"required":["artifact_url"]}}},"tags":[{"name":"tiup","description":"TiUP Publisher service"},{"name":"fileserver","description":"Publisher service for static file server "},{"name":"image","description":"Publisher service for container image"}]}